Refactor CurrencyGroup out from WalletListCurrencyRow
The contents of WalletListCurrencyRow is reused across several places in the AAVE feature.

Refactor the contents of the WalletListCurrencyRow into CurrencyGroup for reuse.
